# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIAA Unified Inference API",
    "version": "1.0.0",
    "description": "Production-grade, hardware-agnostic LLM Inference API following the MIAA Paradigm."
  },
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "summary": "Create Chat Completion",
        "operationId": "createChatCompletion",
        "tags": ["Inference"],
        "security": [{"BearerAuth": []}, {"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "description": "Streamed Response (SSE)"
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "System Healthy",
            "content": {
              "application/json": {
                "example": {"status": "healthy", "layer": "middleware"}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": {"type": "string"},
          "messages": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Message"}
          },
          "stream": {"type": "boolean", "default": false},
          "tools": {"type": "array", "items": {"$ref": "#/components/schemas/Tool"}},
          "tool_choice": {"oneOf": [{"type": "string"}, {"type": "object"}]},
          "thinking": {
            "type": "object",
            "properties": {
              "type": {"type": "string", "enum": ["enabled", "disabled"]},
              "budget_tokens": {"type": "integer"}
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {"type": "string", "enum": ["system", "user", "assistant", "tool"]},
          "content": {"oneOf": [{"type": "string"}, {"type": "array"}]}
        }
      },
      "Tool": {
        "type": "object",
        "properties": {
          "type": {"type": "string", "enum": ["function"]},
          "function": {
            "type": "object",
            "properties": {
              "name": {"type": "string"},
              "description": {"type": "string"},
              "parameters": {"type": "object"}
            }
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "object": {"type": "string", "enum": ["chat.completion"]},
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {"$ref": "#/components/schemas/Message"},
                "finish_reason": {"type": "string"}
              }
            }
          }
        }
      }
    }
  }
}